CDAP-15761 Address comments
